Refactor utils and fix statictic constructor
