+++ src/utils/common.js
@@ -1,8 +1,3 @@
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`
-};
-
 export const getRandomArrayElement = (array) => {
   return array[Math.floor(Math.random() * array.length)];
 };
@@ -29,22 +24,5 @@ export const getTopFilms = (list, characteristic) => {
   } else {
     newlist.sort((item1, item2) => item2[characteristic] - item1[characteristic]);
     return Number(newlist[0][characteristic]) !== 0 ? newlist.slice(0, COUNT_TOP) : false;
-  }
-};
-
-export const createElement = (element) => {
-  const newElement = document.createElement(`div`);
-  newElement.insertAdjacentHTML(RenderPosition.BEFOREEND, element);
-  return newElement.firstChild;
-};
-
-export const render = (container, element, place = RenderPosition.BEFOREEND) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
   }
 };
